requirements.txt
python=3.9.12
geopandas
pytorch
pygeos
tqdm
easydict
pyyaml
transformers
cudatoolkit
trackintel
datasets
